partxn: split transaction into txn and tree type
the parallel txn feature uses a normal transaction that the client
interacts with. but when the db replays intentions, it can do so on a
structure that is more lightweight than the transaction. so this patch
pulls out the guts of the persistent tree, and factors out the
transaction specific bits.

fixes #18

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>