partxn: use intention as rid
This patch stops piggy backing the intention on the rid value, and fixes
up the scenarios for creating unique rids:

rid: this value uniquely identifies a tree delta (a root plus any nodes
creates through tree modifications within a single context). tree deltas are
created in a few different scenarios which need to be taken into acocunt:

  1. An executing txn builds a delta against the last committed state.
  2. The delta produced by replaying an intention record.
  3. Nodes created by loading them from disk in response to a cache miss.

The delta produced in (2) and (3) can be uniquely represented by the
non-negative log address of its after image, or the intention. The after
image produced in (1) must not conflict with any other executing transaction,
or any node accessible via the last committed state.

For (2) and (3), since each log position is by definition unique, any delta
fully represented by a log position can use the log position to uniqely
identify the nodes in the delta. To satisfy (1) let the rid value be a
negative integer, which by defintion, will not conflict with any state
reproduce by intentions from the log. Further, in-flight transactions always
start with a last committed state snapshot, so they are not visible to each
other. Thus, each in-flight transaction can simply use -1 as a rid, however,
we do produce unique negative values to make debugging easier.

To use intention or after image position as the rid? Using the intention is
convenient because we don't need to wait for the after image to be written to
be able to assign the rid values. But when following a pointer during a cache
miss, we follow it to the after image. The intention is currently stored in
the serialization of the after image, so using the intention appears to be
the best choice.

fixes: #25

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>