thrash: cleanup db on exit
this adds more locking to the node, but also destroys the db instance in
thrash. without destrying the instance, there is an assertion failure
thrown in the node. we can't get a backtrace out of it easily so far,
and we still get no error code on exit. it looks like a clean-up issue,
but its only on exit so its probably threads and clean-up all racing.
should make this more robust.

fixes: #65

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>